+++ src/app/core/models/permissions.ts
@@ -1,0 +1,20 @@
+export enum Objects {
+  OFFERS = 'offers',
+}
+
+export enum Permissions {
+  NONE = 'none',
+  ADMINACCESS = 'adminaccess',
+}
+
+export enum UserRoles {
+  ADMIN = 'admin',
+  DEFAULT = 'default',
+}
+
+export class ObjectPermission {
+  object: Objects;
+  permission: Permissions;
+}
+
+