+++ src/app/core/models/user.ts
@@ -1,16 +1,16 @@
-export enum PossibleUserRoles {
-  ADMIN = 'admin',
-  DEFAULT = 'default' 
-}
+import { ObjectPermission, UserRoles } from './permissions';
+
+export type UserStorageData = Pick<User, "token" | "tokenExpirationDate">
 
 export class User {
   constructor(
     public id: number,
     public email: string,
     public name: string,
-    public role: PossibleUserRoles,
+    public role: UserRoles,
     private _token: string,
-    private _tokenExpirationDate: Date
+    private _tokenExpirationDate: Date,
+    private _objectPermissions: ObjectPermission[]
   ) {}
 
   get token() {
@@ -22,5 +22,13 @@ export class User {
 
   get tokenExpirationDate() {
     return this._tokenExpirationDate;
+  }
+
+  public get objectPermissions(): ObjectPermission[] {
+    return this._objectPermissions;
+  }
+
+  public set objectPermissions(value: ObjectPermission[]) {
+    this._objectPermissions = value;
   }
 }